requirements-full.txt
# Core dependencies (required)
crewai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pyyaml

# Database support (optional)
psycopg2-binary
sqlalchemy
alembic

# Redis caching (optional)
redis
hiredis

# Enhanced security (optional)
cryptography
pyjwt

# Web framework for webhooks (optional)
fastapi
uvicorn
pydantic

# Monitoring and metrics (optional)
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation

# Rate limiting (optional)
ratelimit

# Document processing (optional)
pypdf
python-docx
openpyxl

# Testing
pytest
pytest-asyncio
pytest-mock
pytest-cov

# Development tools
black
flake8
mypy
pre-commit